+++ Cargo.toml
@@ -20,6 +20,7 @@ members = [
   "crates/bitnet-kernels",
   "crates/bitnet-inference",
   "crates/bitnet-rope",
+  "crates/bitnet-tokenizers-core",
   "crates/bitnet-tokenizers",
   "crates/bitnet-tokenizer-discovery-core", # SRP: tokenizer path auto-discovery core
   "crates/bitnet-prompt-templates",
@@ -117,6 +118,7 @@ default-members = [
   "crates/bitnet-test-env",
   "crates/bitnet-test-fixtures-core",
   "crates/bitnet-models",
+  "crates/bitnet-tokenizers-core",
   "crates/bitnet-tokenizers",
   "crates/bitnet-tokenizer-discovery-core",
   "crates/bitnet-cpu-activations", # SRP: CPU activation kernels shared microcrate
